Remove duplicate files
It appears that this repo has picked up some files that are duplicate if
treated with case insensitvity. On linux this works okay but on Mac, it
seems like all but one of the duplicates gets obliterated. On nix, the
tarball hash will differ between linux and Mac because the set of files
differ. To make this fix I wrote a script [^1] that:

1. On conflict where the files have the same exact content, delete all
but the oldest file because the change that introduced the newer file
was unnecessary.
2. On conflict where the files differ, delete all but the newest file
since it represents the most recent change.

[^1]: Source code of the script:
      ```python

import sys
import hashlib
import subprocess

def md5(fname):
    hash_md5 = hashlib.md5()
    with open(fname, "rb") as f:
        for chunk in iter(lambda: f.read(4096), b""):
            hash_md5.update(chunk)
    return hash_md5.hexdigest()

def file_git_timestamp(fname):
    return int(
        subprocess.check_output(["git", "log", "-1", "--pretty=%ct", fname])
        .decode("utf8")
        .strip()
    )

files = {}

for file in map(str.rstrip, sys.stdin):
    normalized = file.lower()
    if normalized not in files:
        files[normalized] = []
    files[normalized].append(file)

exacts = []
diffs = []

for k, v in files.items():
    if len(v) >= 2:
        hashes = list(map(lambda f: md5(f), v))
        if all(hashes[0] == x for x in hashes):
            exacts.append(v)
        else:
            diffs.append(v)

for exact in exacts:
    oldest_to_newest = sorted(exact, key=lambda x: file_git_timestamp(x))
    print(f"Keep {oldest_to_newest[0]}")
    for kill in oldest_to_newest[1:]:
        print(f"rm {kill}")
        subprocess.check_call(["git", "rm", kill])

for diff in diffs:
    newest_to_oldest = sorted(diff, key=lambda x: file_git_timestamp(x), reverse=True)
    print(f"Keep {newest_to_oldest[0]}")
    for kill in newest_to_oldest[1:]:
        print(f"rm {kill}")
        subprocess.check_call(["git", "rm", kill])

print(f"{len(exacts)} exact matches")
print(f"{len(diffs)} differing matches")
      ```